moving building into Dockerfile, adding checks for existing files

+++ BuildInitialFactorio.sh
@@ -13,7 +13,22 @@ wget -c https://factorio.com/get-download/stable/headless/linux64 -O factorio-se
 echo "Preparing files..."
 tar xf factorio-server.tar.xz
 rm factorio-server.tar.xz
-mkdir $basedir/factorio/saves
-$basedir/factorio/bin/x64/factorio --create ./factorio/saves/initial.zip
-cp $basedir/factorio/data/server-settings.example.json $basedir/factorio/data/server-settings.json
+if [ -d $basedir/factorio/saves ]
+then
+	echo "skipping save creation"
+	if [ -z "$( ls -A '$basedir/factorio/saves' )" ]
+	then
+		$basedir/factorio/bin/x64/factorio --create ./factorio/saves/initial.zip
+	fi
+else
+	echo "creating save folder and initial save"
+        mkdir $basedir/factorio/saves
+	$basedir/factorio/bin/x64/factorio --create ./factorio/saves/initial.zip
+fi
+if [ -f /$basedir/factorio/data/server-settings.json ]
+then
+	echo "server-settings.json already found"
+else
+	cp $basedir/factorio/data/server-settings.example.json $basedir/factorio/data/server-settings.json
+fi
 echo "Factorio files are now ready, go to /factorio/data to set up your server-settings"
